finish implementing DiGraph

+++ src/Edge_Data.py
@@ -2,7 +2,7 @@ from numpy.core import double
 
 
 class Edge_Data:
-    def __init__(self, src: int, des: int, w: double, info: str, tag: int):
+    def __init__(self, src: int, des: int, w: float, info: str, tag: int):
         self.tag = tag
         self.info = info
         self.w = w
